HIVE-27096: Update dbtxnmgr_showlocks.q.out file (Aman Raj reviewed by Stamatis Zampetakis)
The test runs in a split along with other compaction related tests
thus the order the tests run may change affecting also the output file.

The best solution would be to mask the compacton ID in the output but
this would require other changes that are not yet present in this
branch so for the moment it suffices to update the .q.out file.

Closes #4075