HIVE-25970: Missing messages in HS2 operation logs (Stamatis Zampetakis, reviewed by Zoltan Haindrich)
Revert HIVE-22753 (commit 3b8078ab578add06bcdd36db4ff3b899b9313869).

Preventing the creation of the appender (by returning null) leads to
the message triggering the creation to be lost forever. Moreover, the
memory leak that was observed in HIVE-22753 is no longer feasible with
the fix for HIVE-24590 so the revert is completely safe.

Closes #3048