HIVE-27399: Add support for lateral views with CBO (Stephen Carlin reviewed by Krisztian Kasa, Stamatis Zampetakis)
Previously, the CBO step was skipped when query contained lateral views.

Among the changes that needed to be done:

When the HiveTableFunctionScan is created for a lateral view, all of the inputs from the RelNode are mapped into the new "lateral() function, along with the function containing the lateral view. So for example, if the syntax for the SQL were ...LATERAL VIEW explode(field1) on a table with field1 and field2 where field1 is an array, the resulting HiveTableFunctionScan function would be lateral(explode(field1), field1, field2).

The ASTConverter needed to be changed to handle conversion of the lateral view back into an AST that contains the LATERAL_VIEW node. The non-"lateral view" HiveTableFunctionScan still generates an AST without the LATERAL_VIEW node which is more optimal. The ASTConverter checks to see if the HiveTableFunctionScan contains the a "LATERAL" function in order to keep the LATERAL_VIEW node.

The OpProcFactory code needed to be changed in order to continue to allow a filter optimization. There is pushdown (ppd) logic that was being hit when there was no CBO for lateral views. CBO runs a different path for the ppd logic which is in the OpProcFactory. The lateral view contains 2 node paths that are joined together. The filter can be pushed down the path that contains the base table, but not the path that contains the UDTF. By adding this logic, it prevents a performance regression.

A new rule was also needed for CBO. There is a HiveFilter-HiveTableFunctionScan transpose rule. These RelNodes can be transposed when there is at least one AND condition that can be pushed through the HiveTableFunctionScan. A condition can be pushed through if there are no inputRefs that rely on the output of the UDTF function. We cannot use the Calcite FilterTableScanTransposeRule since that rule uses LogicalFilter and LogicalTableFunctionScan which do not exist in the HiveFilter and HiveTableFunctionScan

The hive.cbo.fallback.strategy property was set to NEVER in some tests involving lateral views to ensure that CBO runs fine without errors.

The q.out changes due to this commit are highlighted below.

In some places, an additional "Select Operator" was added. This should not have any performance impact.

Small changes (UDFToInteger, IF CASE transformation, etc.,) in plans since lateral view queries are now pass through CBO.

In some cases lineage information is lost due to constant folding perfomed in CBO. This is not a regression but something that was always there before, since we don't output lineage for constants (HIVE-27726).

Minor column name changes in plans.

Minor stat changes most likely cause we pass through CBO.

Close apache/hive#4442