HIVE-28968: In CI use dedicated project-local .m2 directory for storing maven artifacts (#5826)
Previously the maven artifacts were stored under .git/.m2 directory. In terms of functionality the exact location (.git/.m2 vs .m2) does not matter. However, putting the artifacts under the .git folder can lead into confusion and false conclusions for people not very familiar with the specificities of the CI.

For instance, someone who is not gonna look into detail on what is present under the .git directory may think that the .git data for the project are huge. Using commands such as `du -h --max-depth=1`, which is also used in some places by the CI, can endorse such false impressions.

Moreover, the .m2 directory is a more common place for storing the maven artifacts in CI/CD pipelines.