HIVE-29008: CI logs do not display information about test execution (#5862)
1. Remove maven's quiet mode (-q flag) and use default logging for running tests
2. Remove Surefire's logger configuration since defaults should work just fine

Maven's INFO messages are useful for diagnosability & information purposes. It is very helpful to have insights about what exactly is currently running in CI and can greatly speed up investigation of issues. 

In the past the INFO messages from the execution of tests were present in CI so this commit restores this functionality and adds a bit more.

Build times are not significantly affected but total storage space for logs increases by 5MB (~20%) after this change.