HIVE-28954: CI fails intermittently due to ephemeral-storage exhaustion (#5840)
Introduce ephemeral storage request and limit for HDB container based on current usage and cluster capacity.

Based on recent runs the HDB container, which executes the tests, consumes 10Gi to 15Gi of ephemeral storage. To ensure that pods are scheduled correctly to the GKE nodes that have the necessary capacity we should add an explicit resource request.

Moreover, to avoid malfunctioning PRs/pods affect the overall health of the cluster we set the resource limit to 20Gi that is reasonably high to permit precommits to run fine and can also guard against accidental changes that may cause disk spikes.