HIVE-28667: Error initializing field trimmer instance when starting HS2 (#5879)
RelFieldTrimmer class is a reflection based visitor that tries to find the appropriate method based on the type (Class) of the argument. The errors occurs while trying to register a method dispatcher for `JdbcSort.class` since there are multiple candidate methods that qualify:
* trimFields(JdbcRel rel, ImmutableBitSet fieldsUsed, Set<RelDataTypeField> extraFields)
* trimFields(Sort sort, ImmutableBitSet fieldsUsed, Set<RelDataTypeField> extraFields)
Given that there is no inheritance relationship between `Sort` and `JdbcRel` it is unclear which method should be used to handle JdbcSort leading to an exception that informs the user/dev about the ambiguous semantics.

It seems that the intention of adding trimFields(JdbcRel...) method was to have a fallback for Jdbc operators that are not handled explicitly by the trimmer but this can never work and creates ambiguity as explained above.
The reason that we get this error only during startup of HS2 and not during compilation is because in the current code it is impossible to call the RelFieldTrimmer code with Jdbc operators other than HiveJdbcConverter and JdbcHiveTableScan.

1. Change trimFields(JdbcRel) to trimFields(JdbcSort) to break the ambiguity and avoid the error. There is no change in behavior since anyways the trimmer cannot be called with a JdbcSort operator.
2. Make CalcitePlanner#HIVE_REL_NODE_CLASSES public mainly to facilitate testing. In fact, it makes sense to expose this information somewhere so why not in CalcitePlanner.