HIVE-29064: Drop usage of repository.apache.org (#5925)
The repository.apache.org is not a general-purpose maven repository so we shouldn't use it to fetch artifacts from there. Unnecessary usage puts stress on the Apache infrastructure and can lead into IP bans and CI failures.
See for more info: https://infra.apache.org/abc/#BL018

In general, the less we rely on non-standard remote repositories the better in many aspects for devs, CI, and consumers that depend on us.